+++ internal/oauthclient/app_test.go
@@ -2,127 +2,105 @@ package oauthclient
 
 import (
 	"context"
-	"fmt"
+	"encoding/json"
+	"net/http"
+	"net/http/httptest"
 	"net/url"
 	"testing"
+	"time"
 
 	"github.com/bluesky-social/indigo/atproto/auth/oauth"
 	"github.com/bluesky-social/indigo/atproto/syntax"
+	"github.com/habitat-network/habitat/internal/pdsclient"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-type testStore struct {
-	sessions     map[string]*oauth.ClientSessionData
-	authRequests map[string]*oauth.AuthRequestData
-}
-
-func newTestStore() *testStore {
-	return &testStore{
-		sessions:     make(map[string]*oauth.ClientSessionData),
-		authRequests: make(map[string]*oauth.AuthRequestData),
-	}
-}
-
-func (s *testStore) GetSession(
-	ctx context.Context,
-	did syntax.DID,
-	sessionID string,
-) (*oauth.ClientSessionData, error) {
-	key := string(did) + "/" + sessionID
-	v, ok := s.sessions[key]
-	if !ok {
-		return nil, fmt.Errorf("session not found: %s", did)
-	}
-	return v, nil
-}
-
-func (s *testStore) SaveSession(ctx context.Context, sess oauth.ClientSessionData) error {
-	key := string(sess.AccountDID) + "/" + sess.SessionID
-	s.sessions[key] = &sess
-	return nil
-}
-
-func (s *testStore) DeleteSession(ctx context.Context, did syntax.DID, sessionID string) error {
-	delete(s.sessions, string(did)+"/"+sessionID)
-	return nil
-}
-
-func (s *testStore) GetAuthRequestInfo(
-	ctx context.Context,
-	state string,
-) (*oauth.AuthRequestData, error) {
-	v, ok := s.authRequests[state]
-	if !ok {
-		return nil, fmt.Errorf("request info not found: %s", state)
-	}
-	return v, nil
-}
-
-func (s *testStore) SaveAuthRequestInfo(ctx context.Context, info oauth.AuthRequestData) error {
-	s.authRequests[info.State] = &info
-	return nil
-}
-
-func (s *testStore) DeleteAuthRequestInfo(ctx context.Context, state string) error {
-	delete(s.authRequests, state)
-	return nil
-}
-
-func testApp(store oauth.ClientAuthStore) *App {
+func testApp(t *testing.T, store oauth.ClientAuthStore) *App {
+	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		require.Equal(t, "POST", r.Method)
+		require.Equal(t, "/oauth/token", r.URL.Path)
+		require.NoError(t, r.ParseForm())
+		require.Equal(t, "abc", r.PostForm.Get("code"))
+		w.Header().Set("Content-Type", "application/json")
+		require.NoError(t, json.NewEncoder(w).Encode(map[string]string{
+			"access_token":  newJwtToken(t, "did:web:test.com", time.Now().Add(time.Hour)),
+			"refresh_token": "refresh-token",
+		}))
+	}))
+	t.Cleanup(testServer.Close)
 	cfg := oauth.NewPublicConfig(
 		"https://example.com/client-metadata.json",
 		"https://example.com/oauth-callback",
 		[]string{"atproto"},
 	)
-	return NewApp(&cfg, store)
+	return NewApp(
+		&cfg,
+		store,
+		WithDirectory(
+			pdsclient.NewDummyDirectory(
+				"pds.com",
+				pdsclient.WithHabitatService(testServer.URL),
+			),
+		),
+	)
 }
 
 func TestNewApp(t *testing.T) {
-	store := newTestStore()
-	app := testApp(store)
+	store := oauth.NewMemStore()
+	app := testApp(t, store)
 	require.NotNil(t, app)
 	assert.Equal(t, "https://example.com/client-metadata.json", app.Config.ClientID)
 }
 
-func TestApp_StartAuthFlow_FailsWithoutServer(t *testing.T) {
-	app := testApp(newTestStore())
-	_, err := app.StartAuthFlow(context.Background(), "did:plc:nonexistent")
-	assert.Error(t, err)
+func TestApp(t *testing.T) {
+	app := testApp(t, oauth.NewMemStore())
+	redirectURI, err := app.StartAuthFlow(context.Background(), "did:web:test.com")
+	require.NoError(t, err)
+	parsedURL, err := url.Parse(redirectURI)
+	require.NoError(t, err)
+	require.Equal(t, "https://example.com/oauth-callback", parsedURL.Query().Get("redirect_uri"))
+	state := parsedURL.Query().Get("state")
+	sess, err := app.ProcessCallback(
+		context.Background(),
+		url.Values{"state": {state}, "code": {"abc"}},
+	)
+	require.NoError(t, err)
+	require.Equal(t, syntax.DID("did:web:test.com"), sess.AccountDID)
 }
 
 func TestApp_GetClient_FailsWithoutSession(t *testing.T) {
-	app := testApp(newTestStore())
+	app := testApp(t, oauth.NewMemStore())
 	_, err := app.GetClient(context.Background(), syntax.DID("did:plc:nonexistent"), "sess1")
 	assert.Error(t, err)
 }
 
 func TestApp_ProcessCallback_InvalidParams(t *testing.T) {
-	app := testApp(newTestStore())
+	app := testApp(t, oauth.NewMemStore())
 	_, err := app.ProcessCallback(context.Background(), url.Values{})
 	assert.Error(t, err)
 }
 
 func TestApp_Logout_Idempotent(t *testing.T) {
-	app := testApp(newTestStore())
+	app := testApp(t, oauth.NewMemStore())
 	err := app.Logout(context.Background(), syntax.DID("did:plc:nonexistent"), "sess1")
 	assert.NoError(t, err)
 }
 
 func TestApp_ProcessCallback_MissingState(t *testing.T) {
-	app := testApp(newTestStore())
+	app := testApp(t, oauth.NewMemStore())
 	_, err := app.ProcessCallback(context.Background(), url.Values{"code": {"abc"}})
 	assert.Error(t, err)
 }
 
 func TestApp_ProcessCallback_MissingCode(t *testing.T) {
-	app := testApp(newTestStore())
+	app := testApp(t, oauth.NewMemStore())
 	_, err := app.ProcessCallback(context.Background(), url.Values{"state": {"abc"}})
 	assert.Error(t, err)
 }
 
 func TestApp_ProcessCallback_StateNotFound(t *testing.T) {
-	app := testApp(newTestStore())
+	app := testApp(t, oauth.NewMemStore())
 	_, err := app.ProcessCallback(context.Background(), url.Values{
 		"state": {"nonexistent"},
 		"code":  {"abc"},
