+++ typescript/internal/src/authManager.ts
@@ -37,6 +37,7 @@ export class AuthManager {
       },
       client_id,
     );
+    this.config.timeout = undefined
     this.serverDomain = serverDomain;
     this.store = create(
       persist<{ authInfo: AuthInfo | undefined }>(
@@ -148,7 +149,7 @@ export class AuthManager {
               this.setAuthState(token);
             }
           })
-          .then(() => {})
+          .then(() => { })
           .finally(() => {
             this.refreshPromise = undefined;
           });
@@ -207,4 +208,4 @@ export class AuthManager {
   }
 }
 
-export class UnauthenticatedError extends Error {}
+export class UnauthenticatedError extends Error { }
